+++ src/main/java/com/sparta/spartaoutsourcing/category/dto/request/CategoryRequestDto.java
@@ -1,0 +1,8 @@
+package com.sparta.spartaoutsourcing.category.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class CategoryRequestDto {
+    private String name;
+}
