--- src/main/java/com/sparta/spartaoutsourcing/domain/category/dto/request/CategoryRequestDto.java
@@ -1,8 +1,0 @@
-package com.sparta.spartaoutsourcing.domain.category.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class CategoryRequestDto {
-    private String name;
-}
