--- src/main/java/com/sparta/spartaoutsourcing/domain/exception/AlreadyDeletedException.java
@@ -1,5 +1,0 @@
-package com.sparta.spartaoutsourcing.domain.exception;
-
-public class AlreadyDeletedException extends RuntimeException{
-    public AlreadyDeletedException(String message) {super(message);}
-}
