--- src/main/java/com/sparta/spartaoutsourcing/domain/exception/AlreadyExistsException.java
@@ -1,5 +1,0 @@
-package com.sparta.spartaoutsourcing.domain.exception;
-
-public class AlreadyExistsException extends RuntimeException {
-    public AlreadyExistsException(String message) { super(message);}
-}
