--- src/main/java/com/sparta/spartaoutsourcing/domain/exception/NotFoundException.java
@@ -1,5 +1,0 @@
-package com.sparta.spartaoutsourcing.domain.exception;
-
-public class NotFoundException extends RuntimeException{
-    public NotFoundException(String message) {super(message);}
-}
