--- src/main/java/com/sparta/spartaoutsourcing/domain/exception/UnauthorizedException.java
@@ -1,8 +1,0 @@
-package com.sparta.spartaoutsourcing.domain.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-public class UnauthorizedException extends RuntimeException {
-    public UnauthorizedException(String message) { super(message);}
-}
