--- src/main/java/com/sparta/spartaoutsourcing/domain/menu/repository/MenuRepository.java
@@ -1,8 +1,0 @@
-package com.sparta.spartaoutsourcing.domain.menu.repository;
-
-import com.sparta.spartaoutsourcing.domain.menu.entity.Menu;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MenuRepository extends JpaRepository<Menu, Long> {
-    boolean existsByNameAndStoreId(String name, Long store_id);
-}
