--- src/main/java/com/sparta/spartaoutsourcing/domain/option/dto/request/MenuOptionRequestDto.java
@@ -1,9 +1,0 @@
-package com.sparta.spartaoutsourcing.domain.option.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class MenuOptionRequestDto {
-    private String name;
-    private Long price;
-}
