+++ src/main/java/com/sparta/spartaoutsourcing/menu/dto/request/MenuRequestDto.java
@@ -1,0 +1,15 @@
+package com.sparta.spartaoutsourcing.menu.dto.request;
+
+import com.sparta.spartaoutsourcing.category.dto.request.CategoryRequestDto;
+import com.sparta.spartaoutsourcing.optionGroup.dto.request.OptionGroupRequestDto;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+public class MenuRequestDto {
+    private String name;
+    private Long price;
+    private CategoryRequestDto category;
+    private List<OptionGroupRequestDto> optionGroups;
+}
