+++ src/main/java/com/sparta/spartaoutsourcing/menu/exception/AlreadyDeletedException.java
@@ -1,0 +1,5 @@
+package com.sparta.spartaoutsourcing.menu.exception;
+
+public class AlreadyDeletedException extends RuntimeException{
+    public AlreadyDeletedException(String message) {super(message);}
+}
