+++ src/main/java/com/sparta/spartaoutsourcing/menu/exception/AlreadyExistsException.java
@@ -1,0 +1,5 @@
+package com.sparta.spartaoutsourcing.menu.exception;
+
+public class AlreadyExistsException extends RuntimeException {
+    public AlreadyExistsException(String message) { super(message);}
+}
