+++ src/main/java/com/sparta/spartaoutsourcing/menu/exception/NotFoundException.java
@@ -1,0 +1,5 @@
+package com.sparta.spartaoutsourcing.menu.exception;
+
+public class NotFoundException extends RuntimeException{
+    public NotFoundException(String message) {super(message);}
+}
