+++ src/main/java/com/sparta/spartaoutsourcing/menu/repository/MenuRepository.java
@@ -1,0 +1,8 @@
+package com.sparta.spartaoutsourcing.menu.repository;
+
+import com.sparta.spartaoutsourcing.menu.entity.Menu;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+    boolean existsByNameAndStoreId(String name, Long store_id);
+}
