+++ src/main/java/com/sparta/spartaoutsourcing/option/dto/request/MenuOptionRequestDto.java
@@ -1,0 +1,9 @@
+package com.sparta.spartaoutsourcing.option.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class MenuOptionRequestDto {
+    private String name;
+    private Long price;
+}
